Change the category code of the percent sign to other in \markdownInput Closes #42 and #79.
This change makes sure that a user of the `hybrid` Lua option or a
malevolent actor can't produce TeX comments in the plain TeX output of
the Markdown package. TeX comments have always been a pain in the
Markdown package, since they allow the user to inadvertently corrupt the
plain TeX output of the package:

    This percent sign % will not only consume this line
    but also the following line. This is because the Lua
    parser (correctly) does not obey newlines in the input.

    The following percent sign will produce a corrupt plain
    TeX outpub, because it will consume the closing brace of
    the `\markdownRendererLink` command:
    [link](https://example.com/some%20pathname%20with%20spaces).

Since 92f217b, we support stripping TeX comments on the level of the Lua
parser using the Lua `texComments` option, which seems to obviate any
possible remaining uses for TeX comments in the plain TeX output of the
Markdown package.