Document that no `eagerCache` makes recursive nesting undefined behavior
Including a nested markdown document fragment when the `eagerCache`
option is disabled causes Markdown to alter a file that TeX is holding
a file descriptor to. After TeX has emerged from the nested document
fragment, it will usually continue from the same position but in a
different file, often typesetting the tail of the nested file twice.
This behavior is undesirable and generally undefined.